refactored e2e tests into separate files to make them easier to maintain
