+++ test/e2e/triggerDumpOpts.test.ts
@@ -1,0 +1,8 @@
+import { dumpFlagTest } from './lib'
+import { TriggerDumpOptions } from '../../src/interfaces/Options'
+
+describe('mysqldump.e2e', () => {
+    describe('trigger dump opts', () => {
+        dumpFlagTest<TriggerDumpOptions>('trigger', 'dropIfExist', /DROP TRIGGER IF EXISTS/)
+    })
+})
